Light Documentation finished
I know this documentation is a bit too redundant, the code itself is easy to understand so having any sort of documentation isn't nessesary. Also Swift-DocC like most things in the swift ecosystem are pain if you're not using a Mac, sure on Linux it's better but still, Mac better for that.